Move database utils sources into the clp namespace.
