+++ components/core/src/clp/GlobalMySQLMetadataDB.cpp
@@ -2,8 +2,8 @@
 
 #include <fmt/core.h>
 
-#include "../database_utils.hpp"
 #include "../type_utils.hpp"
+#include "database_utils.hpp"
 #include "streaming_archive/Constants.hpp"
 
 using std::pair;
